backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
distro==1.9.0
fastapi==0.136.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
openai==2.32.0
orjson==3.11.8
pinecone==8.1.2
pinecone-plugin-assistant==3.0.3
pinecone-plugin-interface==0.0.7
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
requests==2.33.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
tree-sitter==0.25.2
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
